requirements.txt
cyclonedds-nightly == 2023.8.24
pytest == 7.4.2
pytest-timeout == 2.1.0
docker == 6.1.3
requests == 2.29.0
